Include old value to changed.fu.spinbox event
I modified the "changed.fu.spinbox" event to include both current and old values. In my particular scenario I need that to know if the change was an increment or decrement and based on that I can perform a specific action.

This is a breaking change because the event would now send an object instead of a plain value.